update install steps + logic for other tools

+++ tools/kitty.sh
@@ -1,10 +1,12 @@
 #!/bin/bash
-curl -L https://sw.kovidgoyal.net/kitty/installer.sh | sh /dev/stdin \
-    installer=version-0.35.2 launch=n &&
+[[ -d $HOME/.local/kitty.app/ ]] && exit 1
+
+curl -L https://sw.kovidgoyal.net/kitty/installer.sh |
+  sh /dev/stdin installer=version-0.35.2 launch=n || exit 1
 
 mkdir -p $HOME/.config/
 rm -rf $HOME/.config/kitty/
 
-path=$(realpath $0)
-cp -r ${path%/*}/../.config/kitty/ $HOME/.config/
+dir=$(dirname $0)
+cp -r $dir/../.config/kitty/ $HOME/.config/
